requirements.txt
pymongo
paho-mqtt
jupyter
pandas
matplotlib
xlwt
xlrd
tago
